requirements.txt

fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
Pillow>=10.3.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
torcheval>=0.0.7
